+++ test/fixtures/configured/nuxt.config.ts
@@ -3,7 +3,11 @@ import MyModule from '../../../src/module'
 export default defineNuxtConfig({
   modules: [MyModule],
   notify: {
-    strict: false,
+    position: 'bottom-left',
+    duration: 3000,
+    maxToasts: 2,
+    theme: 'dark',
     showIcon: false,
+    strict: false,
   },
 })