+++ vitest.config.ts
@@ -1,0 +1,15 @@
+import { defineVitestConfig } from '@nuxt/test-utils/config'
+
+export default defineVitestConfig({
+  test: {
+    environment: 'nuxt',
+    environmentOptions: {
+      nuxt: {
+        domEnvironment: 'happy-dom',
+      },
+    },
+    include: ['test/**/*.test.ts'],
+    testTimeout: 120_000,
+    hookTimeout: 120_000,
+  },
+})